Fixes #149 Add expression mapping support to transformation editor

+++ transformation/plugins/org.jboss.tools.fuse.transformation.editor/src/main/java/org/jboss/tools/fuse/transformation/editor/internal/VariablesViewer.java
@@ -36,22 +36,23 @@ import org.eclipse.swt.dnd.DND;
 import org.eclipse.swt.dnd.DragSourceAdapter;
 import org.eclipse.swt.dnd.DragSourceEvent;
 import org.eclipse.swt.dnd.Transfer;
+import org.eclipse.swt.events.ControlAdapter;
+import org.eclipse.swt.events.ControlEvent;
 import org.eclipse.swt.events.SelectionAdapter;
 import org.eclipse.swt.events.SelectionEvent;
+import org.eclipse.swt.graphics.Color;
 import org.eclipse.swt.graphics.Image;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.ToolBar;
 import org.eclipse.swt.widgets.ToolItem;
 import org.eclipse.ui.ISharedImages;
 import org.eclipse.ui.PlatformUI;
-import org.jboss.mapper.MapperConfiguration;
-import org.jboss.mapper.MappingOperation;
-import org.jboss.mapper.MappingType;
 import org.jboss.mapper.Variable;
-import org.jboss.mapper.VariableMapping;
 import org.jboss.tools.fuse.transformation.editor.Activator;
-import org.jboss.tools.fuse.transformation.editor.TransformationEditor;
-import org.jboss.tools.fuse.transformation.editor.internal.util.Util;
+import org.jboss.tools.fuse.transformation.editor.internal.util.TransformationConfig;
+import org.jboss.tools.fuse.transformation.editor.internal.util.Util.Colors;
+import org.jboss.tools.fuse.transformation.editor.internal.util.Util.Decorations;
+import org.jboss.tools.fuse.transformation.editor.internal.util.Util.Images;
 
 /**
  *
@@ -61,13 +62,11 @@ public final class VariablesViewer extends Composite {
     final TableViewer tableViewer;
 
     /**
-     * @param editor
-     * @param parent
      * @param config
+     * @param parent
      */
-    public VariablesViewer(final TransformationEditor editor,
-            final Composite parent,
-            final MapperConfiguration config) {
+    public VariablesViewer(final TransformationConfig config,
+                           final Composite parent) {
         super(parent, SWT.NONE);
 
         setLayout(GridLayoutFactory.fillDefaults().create());
@@ -76,33 +75,44 @@ public final class VariablesViewer extends Composite {
         // Create tool bar
         final ToolBar toolBar = new ToolBar(this, SWT.NONE);
         final ToolItem addButton = new ToolItem(toolBar, SWT.PUSH);
-        addButton.setImage(new DecorationOverlayIcon(Util.Images.VARIABLE, Util.Decorations.ADD,
-                IDecoration.TOP_RIGHT).createImage());
+        addButton.setImage(new DecorationOverlayIcon(Images.VARIABLE,
+                                                     Decorations.ADD,
+                                                     IDecoration.TOP_RIGHT).createImage());
         addButton.setToolTipText("Add a new variable");
         final ToolItem deleteButton = new ToolItem(toolBar, SWT.PUSH);
-        deleteButton.setImage(PlatformUI.getWorkbench().getSharedImages()
-                .getImage(ISharedImages.IMG_ETOOL_DELETE));
+        deleteButton.setImage(PlatformUI.getWorkbench()
+                                        .getSharedImages()
+                                        .getImage(ISharedImages.IMG_ETOOL_DELETE));
         deleteButton.setToolTipText("Delete the selected variable(s)");
         deleteButton.setEnabled(false);
 
         // Create table
         tableViewer = new TableViewer(this);
         tableViewer.getTable().setHeaderVisible(true);
-        tableViewer.getTable()
-                .setLayoutData(GridDataFactory.fillDefaults().grab(true, true).create());
+        tableViewer.getTable().setLayoutData(GridDataFactory.fillDefaults()
+                                                            .grab(true, true)
+                                                            .create());
         tableViewer.setComparator(new ViewerComparator());
         tableViewer.setContentProvider(ArrayContentProvider.getInstance());
         // Create columns
+        // TODO add support for changing variable names
         final TableViewerColumn nameColumn = new TableViewerColumn(tableViewer, SWT.NONE);
         nameColumn.getColumn().setText("Name");
+        nameColumn.getColumn().setImage(Images.VARIABLE);
         nameColumn.setLabelProvider(new ColumnLabelProvider() {
 
             @Override
+            public Color getForeground(Object element) {
+                return Colors.VARIABLE;
+            }
+
+            @Override
             public Image getImage(final Object element) {
-                final Image img = Util.Images.VARIABLE;
-                if (editor.mapped((Variable) element))
-                    return new DecorationOverlayIcon(img, Util.Decorations.MAPPED,
-                            IDecoration.BOTTOM_RIGHT).createImage();
+                final Image img = Images.VARIABLE;
+                if (config.mapped((Variable) element))
+                    return new DecorationOverlayIcon(img,
+                                                     Decorations.MAPPED,
+                                                     IDecoration.BOTTOM_RIGHT).createImage();
                 return img;
             }
 
@@ -141,12 +151,11 @@ public final class VariablesViewer extends Composite {
 
             @Override
             protected void setValue(final Object element,
-                    final Object value) {
+                                    final Object value) {
                 Variable variable = (Variable) element;
-                variable.setValue(value.toString());
+                config.setValue(variable, value.toString());
                 try {
-                    editor.save();
-                    editor.updateVariableMappings(variable);
+                    config.save();
                     tableViewer.setInput(config.getVariables());
                 } catch (final Exception e) {
                     Activator.error(e);
@@ -161,7 +170,7 @@ public final class VariablesViewer extends Composite {
                     @Override
                     public void dragStart(final DragSourceEvent event) {
                         LocalSelectionTransfer.getTransfer()
-                                .setSelection(tableViewer.getSelection());
+                                              .setSelection(tableViewer.getSelection());
                     }
                 });
         addButton.addSelectionListener(new SelectionAdapter() {
@@ -169,25 +178,24 @@ public final class VariablesViewer extends Composite {
             @Override
             public void widgetSelected(final SelectionEvent event) {
                 final InputDialog dlg = new InputDialog(getShell(),
-                        "Add Variable",
-                        "Enter a new variable name",
-                        null,
-                        new IInputValidator() {
+                                                        "Add Variable",
+                                                        "Enter a new variable name",
+                                                        null,
+                                                        new IInputValidator() {
 
-                            @Override
-                            public String isValid(final String text) {
-                                for (final Variable variable : config.getVariables()) {
-                                    if (variable.getName().equals(text))
-                                        return "Variable already exists";
-                                }
-                                return null;
-                            }
-                        });
-                if (dlg.open() != Window.OK)
-                    return;
+                    @Override
+                    public String isValid(final String text) {
+                        for (final Variable variable : config.getVariables()) {
+                            if (variable.getName().equals(text))
+                                return "Variable already exists";
+                        }
+                        return null;
+                    }
+                });
+                if (dlg.open() != Window.OK) return;
                 config.addVariable(dlg.getValue(), dlg.getValue());
                 try {
-                    editor.save();
+                    config.save();
                     tableViewer.setInput(config.getVariables());
                 } catch (final Exception e) {
                     Activator.error(e);
@@ -208,42 +216,36 @@ public final class VariablesViewer extends Composite {
                 boolean deleteAll = false;
                 try {
                     for (final Iterator<?> iter =
-                            ((IStructuredSelection) tableViewer.getSelection()).iterator(); iter
-                            .hasNext();) {
+                            ((IStructuredSelection) tableViewer.getSelection()).iterator();
+                         iter.hasNext();) {
                         final Variable variable = (Variable) iter.next();
-                        if (editor.mapped(variable)) {
+                        if (config.mapped(variable)) {
                             if (!deleteAll) {
                                 final MessageDialog dlg =
-                                        new MessageDialog(
-                                                getShell(),
-                                                "Confirm",
-                                                null,
-                                                "Variable \""
-                                                        + variable.getName()
-                                                        + "\" is being used in one or more mappings.\n\n"
-                                                        + "Are you sure you want to delete it?",
-                                                MessageDialog.WARNING,
-                                                new String[] {"Cancel", "Yes", "Yes For All", "No"},
-                                                3);
+                                        new MessageDialog(getShell(),
+                                                          "Confirm",
+                                                          null,
+                                                          "Variable \"" + variable.getName()
+                                                          + "\" is being used in one or more mappings.\n\n"
+                                                          + "Are you sure you want to delete it?",
+                                                          MessageDialog.WARNING,
+                                                          new String[] {
+                                                              "Cancel",
+                                                              "Yes",
+                                                              "Yes For All",
+                                                              "No"
+                                                          },
+                                                          3);
                                 final int button = dlg.open();
-                                if (button == 2)
-                                    deleteAll = true;
-                                else if (button == 3)
-                                    continue;
-                                else if (button < 1)
-                                    return;
-                            }
-                            for (final MappingOperation<?, ?> mapping : config.getMappings()) {
-                                if (mapping.getType() == MappingType.VARIABLE
-                                        && ((VariableMapping) mapping).getSource().getName()
-                                                .equals(variable.getName()))
-                                    editor.unmapVariableMapping(mapping);
+                                if (button == 2) deleteAll = true;
+                                else if (button == 3) continue;
+                                else if (button < 1) return;
                             }
                         }
                         config.removeVariable(variable);
                         tableViewer.remove(variable);
                     }
-                    editor.save();
+                    config.save();
                     tableViewer.setInput(config.getVariables());
                 } catch (final Exception e) {
                     Activator.error(e);
@@ -253,14 +255,16 @@ public final class VariablesViewer extends Composite {
 
         // Populate
         tableViewer.setInput(config.getVariables());
-        nameColumn.getColumn().pack();
-        valueColumn.getColumn().pack();
-    }
 
-    /**
-     *
-     */
-    public void refresh() {
-        tableViewer.refresh();
+        // Expand name and value columns to fill initial width of table each time table is resized
+        tableViewer.getTable().addControlListener(new ControlAdapter() {
+
+            @Override
+            public void controlResized(final ControlEvent event) {
+                final int width = tableViewer.getTable().getSize().x / 2;
+                nameColumn.getColumn().setWidth(width);
+                valueColumn.getColumn().setWidth(width);
+            }
+        });
     }
 }
